Integrate frozen set regressions into existing tests
